Tweak word order
The old sentence was a little odd sounding. And, for whatever reason, my brain skipped over the "into this method" part several times. This wording feels a little better.

You can take or leave this edit if you don't agree that it's better (of course).

Unfortunately the linebreaks necessitated rearranging a lot of text. It it common not to use automatic linewrapping so that the source file is thinner looking?