remove inexplicable use of `{parse: true}`
Within the `Feed#parse` method, we tell the students to add a `{parse:
true}` option when they call `set` to add the `latest_entries` to the
`Feed#entries` collection. However, the entries themselves do not need
to be parsed (and we do not define a `parse` method on the entry model).
In short, this instruction does not make sense.